move InMemoryStore to its own file
